Usability improvements to the spec template
Template should be easier to use with a little more structure
between the sections. Added a comment that people can add None instead
of deleting a section. This should help keep the use of the template
more consistent.

The sub-headings in implementation are renamed to match the existing
fields in launchpad. Should make parsing these easier at some point in
the future.

Tweaked the order so its closer to the current launchpad ordering.
Partly to ease the transition, party because I like the order they have
chosen.

Made a distinction between end-users, deployers and developers. This
has created much confusion in the past.

To reduce the text we force people to delete, trying out adding some
instructions as a comment block.

Change-Id: I28fae5c03da6939471109de7177ac7d2440e7680