Update the template to ask about data model changes
Changes which require modifications to the data model
often have a wider impact on the system.  The community
often has strong opinions on how the data model should
be evolved, from both a functional and performance
perspective, and disagreements and/or rework late in the
implemetaion stage are vrey disruptive and tiem consuming.
It is therefore important to capture and gain agreement
as early as possible on any proposed changes to the data
model.

Change-Id: I071043c7ce29bb8a47a9da9fb458befe91f6a609