Update the template to capture additional deployer impact info
Add prompts to help a deployer understand if the change will
take immediate effect, or is something that has to be explicitly
configured into the system.  Generally a change that has
immediate effect would be of more concerns for both updates
and folks on the CI path than a change that is in effect dormant
until configured.

This also of course prompts for the BP submitter to also
consider the same issue, which is a topic that should be addressed
at design time.

Change-Id: I7e78ffcaec800bbfe09135d90069158dc6e58923