Adding section for security impact
The best way to improve security throughout OpenStack is to be more aware
of the security impact that design decisions have.  I think that adding a
section that focuses on security will help encourage people to think of this
during the design process.

This idea originated in an email thread on the openstack-security mailing
list:
http://lists.openstack.org/pipermail/openstack-security/2014-March/001351.html

Change-Id: I7945f7c99387aedabbce057130f3589831f77f1c