Allow DB migration backports to Icehouse
Just as we did at the beginning of the Havana and Icehouse dev cycles,
we need to reserve a range of DB migrations as the first DB change in
Juno. This will allow a range to be used for migration backports to
Icehouse if needed.

This is a straight forward blueprint that has been implemented for the
last two releases (Havana and Icehouse), so we can use this as a trial
run for our new blueprint review process before we open up all other
spec reviews.

Since this is the first Juno spec, re-add the Juno specs directory to
index.rst.

Related to blueprint backportable-db-migrations-juno.

Change-Id: I791c22b54ac8cd533d6fdc03ea93f12d4a907b23