Nova Pagination Sorting Enhancements
Currently, the pagination support for Nova does not allow the caller to
specify the sort order and direction of the data set. This blueprint
enhances the pagination support for the /servers and /servers/detail
APIs so that multiple sort keys and sort directions can be supplied on
the request.

bp nova-pagination

Change-Id: I51aa0799e42f2ad253489d8ba6a1e05a344eb145