Better Support for Multiple Networks
Since nova-network is staying around, it needs a few updates for
multiple networks. There are various settings that are automatically
determined or set via flags, that should be explicitly set per network.
This spec is about adding a few options to the networks table and
converting the network manager and linux_net code to support multiple
networks.

Change-Id: Id57fcdfc4110086af06a881c9a377a824f36d213