make title check more explicit on failure
the title check used to compare for length of arrays, which meant
that if it failed it wasn't really clear why. Given that we know
the 7 titles we believe should exist (and we test for them later)
we can instead just do a direct comparison. This has the advantage
of making it much easier to figure out what they are missing in
their specification.

Change-Id: Ie913f00cfb07a153fc94c8e5b2225174da3674dd