Get expected titles from template
This change removes the data block describing which sections +
subsections must be present. Instead this information is retrieved from
the template. This means the test won't need to be updated if the
template changes.

It also means the test can be shared byte-for-byte with other projects
who have different templates, avoiding the copy+paste situation which
has emerged.

Change-Id: I4d0ea6e931876f70618912759d2de8e37b8b87c6