Change compute updates from periodic to on demand
Currently, all compute nodes update status info in the DB on a periodic
basis (the period is currently 60 seconds). Given that the status of
the node only changes at specific points (manyline image
creation/destruction) this leads to significant DB overhead on a large
system. This BP changes the update mechanism to only update the DB when
a node state changes, specifically at node startup, instance creation
and instance destruction.

Change-Id: I9a53ba31619bab1b25d25b533beba8c86606637c
Implements: blueprint on-demand-compute-update