Add v2-on-v3-api spec
Note that this change includes an addition to the doc
index which is needed to bootstrap the approved directory
for Kilo. Without the change a sphinx warning about
the new spec rst file not being included anywhere causes the
build to fail. And applying the change with an empty approved
directory the pattern match fails and the build fails

Previously-approved: juno
Part of blueprint v2-on-v3-api

Change-Id: Ib8b259f69c9072520acecba167b7b31707bcadaa