Virt driver large page allocation for guest RAM
This feature aims to improve the libvirt driver so
that it can use large pages for backing the guest
RAM allocation. This will improve the performance of
guest workloads by increasing TLB cache efficiency.
It will ensure that the guest has 100% dedicated RAM
that will never be swapped out.

Previously approved: juno
Change-Id: Ie3e3337241a514723d2c85b2186ee15ee8d57fff