Deprecate `.image.glance` in favor of glanceclient
Nova currently maintains a wrapper on top of glanceclient which is both
old and ugly. This spec proposes getting rid of that code in favor of
using just glanceclient in order to reduce the code needed to be
maintained by the nova team, allow the glance team to move forward more
easily and allow nova for better supporting glance's API v2.

Change-Id: I1a86fcf3f2321f2b16b8787faecd099d80c1e3c4