Add notes on the level of detail required
As discussed at the summit, we want to be clear on the level of detail
required in nova-specs.

Previously we had required so much detail that you ended up having to
write all the code upfront, defeating some of the original aims of
getting people to think about their designs upfront.

Change-Id: I03b32b4a98a0d5764ed3f3e8c80236e5dcc61055