Add Spec Backlog
This is copying what has been done for keystone:
If546724fd535db7753a372389c3f90f3b060d9bc

We add a place for people to add rough ideas, and a place to put specs
that are no longer going to be implemented by the proposer.

Includes updating the README to tell people where the backlog specs will
live.

We need to add a backlog spec before we can complete this change, this
is just a placeholder to advertise the backlog idea.

Change-Id: I667acfa422ef78bdc80efb279911091116b1c1f0