Validate database migrations and model
Database migrations and the database model are managed independently in
Nova. When a new database migration is added, changes are often needed
to the database model and vice versa. However, there are no tests or
checks that the database model matches the results of the database
migrations leaving them to often drift apart.

Change-Id: I3671faa62c22179addaf0f24f918eaffa7dbe30f