Adds lock-free-quota-management spec
This was proposed and considered last cycle but failed to get approval
in time for the spec freeze. We're still very much interested in working
on this effort, however, and based on my research with Peter Boros from
Percona, the quota management database layer in Nova is the number one
bottleneck and cause for database-layer scalability issues. This spec
outlines one strategy to avoid the use of our retry_on_deadlock
current behaviour and hopefully get us towards a more concurrent method
of doing quota reservations.

Change-Id: I9af4fafd5821b657a11b1d02bf8f1e846852b563
Blueprint: lock-free-quota-management