Update the testing part of the Template
Because of the way the testing part of the specs template is written
you basically get one of 3 throw away answers in the spec.

1) we'll have unit tests
2) we'll add tempest tests
3) we'll have 3rd party testing

However, the specs *rarely* actually talk about what's actually
important to test in the feature. Which means those are all kind of
throw away statements.

A much more useful piece of information is to ask what are the
critical scenarios to test for a feature. What's really important that
works. What's really important that it doesn't break something
else. What are some interesting edge cases we should be careful about.

This is an initial stab at that paragraph change. We should probably
wordsmith on it though.

Change-Id: I83a146a4561fac50fdd3bd961d16ff32cccbf4b9