New VIF type to allow routing VM data instead of bridging it
We propose to add a new VIF type, VIF_TYPE_TAP, whose meaning is that
the host side of a VNIC to a VM is - at least initially - a simple TAP
interface that is not plugged into a bridge or vSwitch.  A Neutron
agent can continue the process of deciding how to handle data from
that interface, and how to deliver data to it - but this is beyond the
scope of the initial TAP interface setup that Nova needs to provide.

Previously-approved: kilo
Change-Id: I3c5371b865eadc458ce25ba85953d1ddecab2506