process: add new section History
History are used to make easy users to follow the
evolution of the spec.

The spec can differ with the time, and the design can be updated
it should be important for users who use the spec has a documentation
to see those changes without to use a third tool.

The initial proposal is to add an entry for each update provided to
the spec (even fixing a typo) but we can restrain it only for changes
in the design or new contributor refered...

Change-Id: I13b8e09d5d7848ee4657214a86d2b48bf1b05e2a