Cleanup 'scheduled_at' from instances table
The 'scheduled_at' field is a part of the Nova instances.
Currently, the 'scheduled_at' field is no longer updated
by the Nova scheduler, instead the 'launched_at' field is
updated for the same purpose. The 'scheduled_at' column is
now redundant and should be removed as a part of DB clean up.

Change-Id: I7ccf1d225c983b69c05f2415094071bdafd63b85
Blueprint: cleanup-scheduled-at