Add a Distinct Exception for Exceeding Max Retries
There are two separate and very different situations which raise a
NoValidHost exception when attempting to spawn a new VM, which is
confusing for operators who need to determine why a VM failed to spawn.
This proposes to add a new exception class that will make the two
failure modes more obvious to the operators.

Blueprint: no-valid-host-reporting

Change-Id: I1d6784249ff5c99d1a370d046d9078bb250e0974