Add ability to support discard/unmap/trim for Cinder backend
Currently, libvirt/qemu has support for a discard option when
attaching a volume to an instance. With this feature, the
unmap/trim command can be sent from guest to the physical
storage device. A cinder back-end will report a connection
capability that Nova will use in attaching a volume.

Change-Id: I4acb9dbb3c239c1b6b94dca3d2ab438296685a33
Implements: bp cinder-backend-report-discard