Re-propose spec for refresh-quotas-usage
This was approved for liberty but code was never submitted.  I emailed
Romain and he said he wouldn't have time to work on it in Mitaka.
However, this was something that many operators expressed interest in
before so it seems valuable and straight-forward to try again on this in
Mitaka.

There are two changes from the original spec:

1. --user is optional with the quota_usage_refresh command.

2. The testing section is updated to point out that unit and functional
   tests would be provided.

Previously-approved: liberty

Spec for blueprint refresh-quotas-usage

Originally authored by: Romain Hardouin <romain.hardouin@cloudwatt.com>

Change-Id: Idbb571c2b953d8cf9e9b98496bb584d6ca931272