Add thread policy note to virt-driver-cpu-pinning
CPU thread policy manipulation was ultimately not implemented as part of
the virt-driver-cpu-pinning work done in Kilo and now exists (or at
least did for Liberty, pending re-proposal for Mitaka) as a separate
proposal.

This patch updates the original specification to note this and add a
link to the newer blueprint in the hope of removing confusion.

Change-Id: If014840ac5a79dfaeffc56dd56d9627bc0208324