Centralize Config Options
Nova has around 800 config options [1]. Those config options are the
interface to the cloud operators. Unfortunately they often lack a
good documentation which
* explains their impact,
* shows their interdependency to other config options and
* explains which of the Nova services they influence.
This cloud operator interface needs to be consolidated and one way of
doing this is, to move the config options from their declaration in
multiple modules to a few centrally managed modules. These centrally
managed modules should also provide the bigger picture of the
configuration surface we provide. This got already discussed on the
ML [2].

[1] see the "nova.flagmappings" file which get generated in the
    "openstack-manuals" project for the "configuration reference" 
    manual.
[2]
http://lists.openstack.org/pipermail/openstack-dev/2015-July/070306.html

Implements blueprint centralize-config-options

Change-Id: I182e368dd2896aad643f539e7bb2493daf5ead5c