Virt image properties boot time override
This extends the Nova boot API so that it is possible to override any
image metadata properties at instance creation time. This avoids the
need to upload the same image to glance multiple times in order to
set slightly different properties for special use cases.

APIImpact
Previously-approved: Mitaka
Blueprint: virt-image-props-boot-override
Change-Id: I7dca14fbc74fa5b6f4d29b72be1acd48f9b48fa0