template: remove "Project priority" section of spec
The "Project priority" section in the spec template is a source
of persistent confusion and mis-use. A large portion of spec
authors believe that it is a section to fill with their own
opinions on why their spec is important. In fact, pretty much
all specs should put the words "None" in this section. This is
particularly true earlier in the cycle, since the actual project
priorities do not get finalized / agreed on until the design
summit. Even once the priorities are agreed, this section of
the spec is still fairly pointless, since we've moved to tracking
priority specs in a separate etherpad. eg

  https://etherpad.openstack.org/p/mitaka-nova-priorities-tracking

Just remove this source of confusion since it doesn't serve any
particularly useful purpose anymore, given the external tracking
done.

Change-Id: I808227fe71210abc64270f53ef3d3d62f176ff24