Move thread policy features between specs
The thread policy aspects of the virt-driver-cpu-pinning feature
are incorrect. There were not implemented for Kilo and shouldn't be
included in this doc. While there was a note stating as much, it makes
more sense to remove the documentation from the CPU pinning spec
and move it into the (renamed) CPU thread pinning spec.

Change-Id: I28980de107ae8e5b8b8c8394bf7b007ff028e855