Neutron Routed Networks
In Neutron, there is a effort to support routed provider networks.  A
routed network in this context, is a physical network infrastructure
that implements scaled networks by routing instead of large L2
broadcast domains.  For example, deployers may have routers at each
top-of-rack.  Instead of a single VLAN covering the deployment, each
rack would have its own VLAN and the routers will provide reachability
to the rest of the racks over L3.  Operators want Neutron to model
this like a single network.  This has implications for Nova
scheduling.

Depends-On: I1a71e83d9fb3d13306979934fbd11faf80091ab8
Change-Id: I8f5523930d360a9183ab111e015ef4b052e0968f
Related-Bug: #1458890