Add "disable local disks" option to flavors
This feature adds the possibility to define a flavor,
that does not create any ephemeral disks locally on the hypervisor.
If the new flavor key 'local_disks' is set to 'False', users are
"forced" to create a bootable volume to launch an instance from
that flavor.

APIImpact
DocImpact

Change-Id: I52b8d473d86ce471e61a2c3505f6c274ad0643b8
Implements: blueprint flavor-root-disk-none