Get Me a Network
Neutron added the ``auto-allocate-topology`` API in Mitaka [1]_. Nova intends
on using this API to automatically allocate a private network for a project
when creating server instance if no specific network is provided in the create
request and none are available to the project. This should also bring some
feature parity for the server create flow between Neutron and nova-network.

Spec for blueprint get-me-a-network

APIImpact

Change-Id: If34d760deb5fc1c580b4c84d0a569b532bd1385f