Re-propose spec for refresh-quotas-usage
This was approved for Liberty and Mitaka but code was never submitted.
This was something that many operators expressed interest in before
so it seems valuable and straight-forward to try again on this in Newton.

There are two changes from the original Liberty spec:
1. --user is optional with the quota_usage_refresh command.
2. The testing section is updated to point out that unit and functional tests would be provided.

Previously-approved: Mitaka Spec for blueprint refresh-quotas-usage

Originally authored by: Romain Hardouin <romain.hardouin@cloudwatt.com>

Change-Id: Ie71cc6836b1a291c07df3db913ea8655420819ba