Adds Nova discoverable policy CLI spec
https://blueprints.launchpad.net/nova/+spec/discoverable-policy-cli

Not all users have the same set of privileges. Not all API endpoints
are enabled. Because of this, adding an nova-manage CLI subcommand that
lists the operations an user is allowed to use can be helpful. This spec
describes phase one of a two phased implementation. The first phase
will prove out the idea using a raw representation that will be
improved in the second phase.

Co-Authored-By: Chris Dent <cdent@anticdent.org>

Implements: blueprint discoverable-policy-cli

Change-Id: I973f1e0545c73d66ef17bed77c6dda5de007cb1e