Embed policy defaults in code
Given a deployed policy file it is not trivial to determine how much it
differs from the defaults that Nova expects. This is due to the provided
sample file not having an exhaustive list of all configurable policies,
and the expected defaults are not contained in the Nova codebase.

Given an authenticated request context it is not possible to determine
which policies will pass. This is because policy checks are ad hoc
throughout the code with no central registry of all possible checks.

bp policy-in-code
Change-Id: I4e2209f85725535a724e83c576dc07e6df05f439