Re-propose resource-providers for Newton
This blueprint partially addresses the problem of Nova assuming all
resources are provided by a single compute node by introducing a new
concept -- a resource provider -- that will allow Nova to accurately
track and reserve resources regardless of whether the resource is being
exposed by a single compute node, some shared resource pool or an
external resource-providing service of some sort.

The spec has been annotated to reflect which work has been completed and
which remains to be done in Newton.

Change-Id: Id0c030159fa741a5eb51d52332d6fdbce05000ae
Previously-approved: Mitaka