Live Migration of Rescued Instances
When an instance is rescued the libvirt driver creates a file in the.
instance directory that contains the original instance xml. This file
is used when unrescuing the instance to revert to the original disk
configuration. However, this file is not necessary to restore the
instance's original configuration. Furthermore the reliance on this
file means we currently cannot live migrate a rescued instance.  Thus
it is proposed that we do not create this file during the rescue
operation and rebuild the instance's original xml file from the
database when we unrescue it.

APIImpact
Change-Id: Id3a055f00c3f8ef006911e2523c8d284f449b524