libvirt: support for virtozzo storage volume driver
The purpose of this spec is to add a support for volumes hosted
by Virtuozzo storage.

Virtuozzo Storage is a fault-tolerant distributed storage system
optimized for virtualization workloads. From client's point of
view it is a remote file system storage similar to NFS or GlusterFS.

A third-party vzstorage-ci system is running testing jobs on nova
changes.

CI wiki page:
https://wiki.openstack.org/wiki/ThirdPartySystems/Virtuozzo_Storage_CI

blueprint libvirt-vzstorage-volume-support

Previously-approved: Mitaka

Change-Id: Ide2c952c577b36ad1b7f72012336f300f2f3331c