Fix additional mistakes in the NUMA spec document
The 'hw:numa_mempolicy' parameter was never implemented and will likely
never be, however, it's located in the 'implemented' folder suggesting
otherwise. Seeing as specs are as much a reference as anything, this
invalid information should not be retained. Correct this, taking the
opportunity to fix some other typo, line wrapping and general
formatting.

Change-Id: I4aca0073f1fa26ff0c3a34407370f6ba6d916879