Create specs directory for Ocata
The deadline for non-priority specs for Newton is almost here,
and as always there will be a large number of specs that do
not get approved. Any that remain in the Newton directory will
eventually get force abandoned. Opening up the Ocata directory
allows submitters who wish to re-target their effort to the
subsequent release, to move their spec across to the new
directory straight away enabling ongoing visibility & review.

Change-Id: I732576bb3c0e60ec130cdf73647fbc2f234905f7
Signed-off-by: Daniel P. Berrange <berrange@redhat.com>