Add a script to move completed specs to implemented directory
This adds a script and tox target to automatically process and
move approved specs for a given release into the related implemented
directory for the same release. It uses launchpadlib to check the
status on the spec (blueprint) and only moves those that are
complete.

For example, to run this:

tox -r -e move-implemented-specs -- -n -v newton

Change-Id: Ib431f62101b90abecce86f60ba7acbba11e09533