Fix spec names to match blueprint names
The filename of the spec should match the blueprint
in launchpad, this is what the move-implemented-specs
script works with and needs, and our template/review process
should enforce. This change fixes the newton specs that
had busted names and updates the redirects file to preserve
any links to the old names.

You can test this by running:

tox -r -e move-implemented-specs -- -n -v newton

And there won't be any warnings at the end.

Change-Id: I806bdebde1c5bf8e48dfdbf4813342a9c4de5ad3