Add "disable local disks" option to flavors
This feature adds the possibility to define a flavor,
that does not create any ephemeral disks locally on the hypervisor.
If the new flavor key 'local_disks' is set to 'False', users are
"forced" to create a bootable volume to launch an instance from
that flavor.

Change-Id: I428b26a88bb6df500ad4a3daab158875e59050be
Previously-Approved: Newton
Implements: blueprint flavor-root-disk-none