Proposal for cells v2 scheduling interaction
The scheduler should be called earlier in the build process so that an
instance can be created in the proper cell.

This spec has been previously approved in Newton but minor modifications
have been made in this version. Details have been added about the need
for a new conductor method.

blueprint cells-scheduling-interaction

Change-Id: I9a3cf1bec13143771231aa3602c0a955c847ca47