Libvirt: Use the virtlogd deamon
If the *serial console* feature is enabled on a compute node with
``[serial_console].enabled = True`` it deactivates the logging of the
boot messages. From a REST API perspective, this means that the two
APIs ``os-getConsoleOutput`` and ``os-getSerialConsole`` are mutually
exclusive. Both APIs can be valuable for cloud operators in the case
when something goes wrong during the launch of an instance. This
blueprint wants to lift the XOR relationship between those two REST
APIs.

Bug 1599214 in Qemu made it necessary to block the usage of "virtlogd"
in Newton. Qemu 2.7.0 fixes the aforementioned issue which makes a
re-proposal of this formerly approved spec possible. This change updates
the Qemu version requirement.

Previously-approved: Newton
blueprint libvirt-virtlogd
Change-Id: Ia97c3fd6023687ee4b7903a0f8b30716677b10dd