Enable cold migration with target host
The target host can be specified when cold migrating a server.
And a new request body field called ``force`` (defaulted to False)
is added in order not to call the scheduler when providing a
destination. Setting it to 'True' makes the cold migrate operation
bypass the scheduler.

APIImpact
Change-Id: I57f3cb045a5f050931dc6c0fe9786ddd7dd0a2d7
blueprint cold-migration-with-target-ocata