Send service token along with user tokens
Some operations in Nova could take a long time to complete. During this
time user token associated with this request could expire. When Nova tries
to communicate with other services using the same user token, Keystone fails
to validate the request due to expired token.

Service token will be passed along with user token to communicate with
services when dealing with long running tasks like live migration.
Keystone middlewere trusts that the service got the user token when
it was valid, don't check the expiry date of cert.

Change-Id: I7f5807ed6b998a0d88169dbbf057e22fe2cd9167