Provide error codes for placement API
This blueprint describes a framework for providing additional
information in case of error when making requests of the placement API.
This information can be used on the client side to distinguish what
caused an error in the event of the same HTTP response code.

This is a re-proposal of a spec originally proposed to Queens. The main
changes are:

* It removes the requirement to implement the help link, suggesting
  we do that later.
* It adjusts the emphasis to being development of a framework rather
  than a complete overhaul of existing responses.
* Add links to:
    * a WIP of how to do it
    * placement api-ref
* Makes some comments in the Alternatives section about how we may
  wish to address microversions.

blueprint placement-api-error-handling
Co-Authored-By: Chris Dent <cdent@anticdent.org>
Change-Id: Id644f9f509ab382afe155cca0babf08a53d66338