Record reality of placement API usage from n-sched
The resource-providers-scheduler-db-filters spec previously contained wording
that intimated that the scheduler would begin to use the placement API for
scheduling decisions as soon as some nova-compute workers were upgraded and
began reporting inventory and allocation records to the placement API. This is
not how Ocata was released. Instead, the placement API is only queried from the
scheduler once all nova-compute workers have been upgraded to Ocata. This patch
simply records that new reality.

Change-Id: I9b9e3855b8479dd22d714149889793224909832f