tox: Build docs with Python 2.7
On systems that use Python 3 as default, the 'docs' target fails due to
a non-Python 3 dependency (feedformatter, required by yasfb). It's not
even clear if yasbf is still relevant, given the uselessness of the
content published online [1]. However, it's easier to just build the
docs with Python 2 than go on a wild goose chase looking for more info.

[1] http://specs.openstack.org/openstack/nova-specs/rss.xml

Change-Id: I129f635bc2d060f9da9bfc8b43652be7c10e9200