Re-add templates
In commit a67d39ba, we removed templates from both the approved and
implemented folders of 'specs/{release}'. However, git tracks files, not
folders, meaning the 'implemented' folders for both Ocata and Pike,
which are currently empty, are no longer tracked and the glob patterns
used in 'doc/source/specs/{release}/index.rst cause warnings.

These templates are re-added to fix the immediate issue. However, we
still don't want to keep these around once the release has been made. As
such, we add an addition step to the tool for migrating implemented
specs to remove these templates. This should automate the process in the
future.

Change-Id: Iaae10e180845b5073e1e13a4df1050b8b98d03d7