Use uuids in services and os-hypervisors APIs (spec)
To work with services and hypervisors (compute nodes) in the compute REST API
we currently expose and take primary key IDs. In a multi-cell
deployment, these IDs are not unique. This spec proposes exposing a uuid for
services and hypervisors in the REST API to uniquely identify a resource
regardless of which cell it is in.

APIImpact

Spec for blueprint service-hyper-uuid-in-api

Change-Id: If0a3d93bb3d41c4637caf46722e66c589cabcab4