Remove Location header from createImage and createBackup response (spec)
The `createImage` and `createBackup` APIs return a `Location` header in the
response which is a URL pointing at what is most likely an internal Glance API
server, and is therefore not accessible to the end user. The URL is also not
versioned for the Image service, so unless the cloud provides an alias, the URL
is wrong anyway. This spec proposes to replace the Location header with a simple
response body json dict that contains the image_id for the snapshot image.

APIImpact

Spec for blueprint remove-create-image-location-header-response

Change-Id: I976c0f6775d3020028e3e4cc0ed3bc67a70c421d