Amend spec for placement project/user association
This changes from using a GET of /allocations to a GET of /usages for
getting a summary of resource usage from placement and adds required
properties: project_id and user_id to the /allocations PUT method
schema.

Other comments from the original review are also addressed such as
mentioning that a new microversion will be needed, defining better
table indexes, and using monospaced text for refer to table names,
model names, and REST resource names.

Spec for blueprint placement-project-user

Change-Id: I6a428db4db86c7fc9e1cccb265b42a30b23885a1