Clarify wording in Searchlight spec - error conditions
There was some confusing / conflicting wording in the error
conditions part of the spec (do we log error or warning?). This
clarifies that to be consistent about saying we just log a warning
if configured to use Searchlight but it's not available (which is
what we do if Placement isn't available).

Related to blueprint list-instances-using-searchlight

Change-Id: I1331d685e82f100a1965bd835e89147c10cbcda1