Add a spec for minimal cache headers in placement
rfc 7232 section-2.2 says that a web service should send a
last-modified header for any representation for a last-modified
time can reasonable be deduced. The placement service does not
currently do this. Including it will make it a better HTTP citizen
and also provide useful metadata in the response. If a
last-modified header is added then it is also necessary to add a
cache-control header with a value of "no-cache" to insure that clients
and proxies are not inclined to cache representations provided by the
placement service.

Change-Id: Icd5a2682aa49baa27ec61b93c491a66f000dd34d
Blueprint: placement-cache-headers