Return Selection Objects
In Queens, we are going to be returning a lot more information from the
Scheduler to the Conductor. If we return it as a series of nested
dictionaries and lists, it will be fragile and confusing. It would be
far better to create an object to hold this data and make it accessible
in a simpler and documented way.

Blueprint: return-selection-objects

Change-Id: I7e94525ba9ee22885ea18a752a416024c36b4a40