Add a spec for POST /allocations in placement
For the reasons described in the spec it would be nice to be able to
POST to /allocations to set the allocations for multiple consumers in
one request.

This spec leaves open several questions about body representation and
behavior that ought to be resolved as part of the review process.

Note that the JSON examples are handmade so may suffer from errors which
will be cleared up in the implementation

Change-Id: Ib4a3970a5d0e475c9b0a7fcb5f51e1ed3662befa
Blueprint: post-allocations