Return Alternate Hosts
Sometimes when a request to build or resize a VM is attempted, the build
can fail for a variety of different reasons. At recent PTG and Forum
events we discussed a request from operators to still be able to retry
failed builds. With the cells V2 design, though, the communication from
the cell conductor to start the retry process will no longer be
possible, as cells cannot call up to the API level. We propose to have
the scheduler return some alternate hosts along with the selected host.
This was desired because in the event of a failed build or resize,
another host could be tried without having to go through the entire
scheduling process again.

Blueprint: return-alternate-hosts

Change-Id: I3b6fcfa67425c0865e174c5a921fbfa9596c76d9