Support traits in the Ironic driver
We need scheduling that is more granular than resource classes. Traits
are the best way to do that. This talks about how we integrate ironic
nodes having associated traits into a Nova boot.

This spec depends on changes in Ironic to allow operators to set
traits on ironic nodes, alongside the existing ability to add a resource
class for each node.

blueprint ironic-driver-traits

Change-Id: Ice5468aaefc2f6d0ae8f6b0e86f2096f45b75919