Propose counting quota usage from placement and API database
In Pike, we re-architected the quota system to count actual resource
usage instead of using reservations and tracking quota usages in a
separate database table. We're counting resources like instances, CPU,
and RAM by querying each cell database and aggregating the results per
project and per user. This approach is susceptible to undesirable
behavior if a cell becomes unavailable. If a cell becomes unavailable,
resources in its database cannot be counted and will not be included in
resource usage until the cell returns. Cells could become unavailable
if an operator is performing maintenance on a cell or if a cell
database is experiencing problems and we cannot connect to it.

We can make resource usage counting for quotas resilient to temporary
cell outages by querying placement and the API database for resource
usage instead of reading separate cell databases.

Spec for blueprint count-quota-usage-from-placement

Change-Id: Ib86a7994d7a577b8e6a2dd32f4fcf12b7f788af8