Add serial numbers for local disks
It is possible to assign a serial number to virtual hard disks which
can be queried by the guest operating system. We currently do this for
attached volumes, but we don't do it for local disks: local root
disks, ephemerals, or swap disks.

The primary user-facing purpose of this feature is to expose a
persistent serial number for local disks to guests. A useful
consequence of this feature will be to add a uuid to block device
mappings, which has been frequently requested for other features.

Change-Id: I0ffa5ed1738804b636a0024a42e1243b2d2a043d