Remove the py27 tox target
With the move to zuulv3, the openstack-tox-py27 job definition
ignores changes to docs and .rst files, which is basically
all of nova-specs, so we aren't getting our spec linters run
as part of the py27 target anymore.

This moves them into the pep8 target where they should probably
live as linter tests anyway.

Change-Id: Ic87e79db6a8881398d670148b609e0b81658dfd1