Amend the migration paging spec for uuid in server migrations response
During the review of I7e01f95d7173d9217f76e838b3ea71555151ef56
we made the decision to also add the migration uuid to the server
migrations GET responses to be consistent with the os-migrations
change to add the uuid to the response. This change simply amends
the spec to mention that change as well.

Part of blueprint add-pagination-and-change-since-for-migration-list

Change-Id: Icdd598d34748240a3ed9dade561f8f34857adbc3