Allow abort live migrations in queued status
The functionality of abort live migration was added in microversion 2.24,
and currently only migrations in ``running`` status are allowed to be
aborted.

While the number of live migration requests could be greater than the max
concurrent live migration configuration ``max_concurrent_live_migrations``,
and there will be migrations wait in queue. The migrations could remain in
``queued`` status for a very long time depend on the queue length and the
processing speed.

Admins may want to abort migrations in queue due to time consumption
considerations etc. It will be unreasonable to make admins wait until
the status turn to ``running`` before they can be aborted.

This spec adds support to allow abort live migrations in ``queued``
status.

APIImpact

Change-Id: I61881e2dc079412631c59a70567dd4e09eb257ca
Implement: blueprint abort-live-migration-in-queued-status