Add generation support in aggregate association
As resource provider handling has evolved, it has become clear that
virt drivers may wish to manage a complex collection of interrelated
resource providers, resource class, traits and aggregates. Some of
those entities may need to be managed by multiple threads of
processing, including different nova-compute processes.

The `generation` concept in the placement service provides a tool
for managing these kinds of concurrent operations: a write API
request that includes a generation has that generation compared against
the value on the server. If there is a match the request can go
ahead, otherwise a ``409 Conflict`` response is returned.

This functionality is not present when associating or disassociating
a resource provider with one or more aggregate uuids. For virt drivers
that wish to manage aggregate associations safely, this is necessary.
This spec describes the changes necessary to add the functionality.

Change-Id: I18b0fd3a0b2f4ddeb74182d3e79fead659894001
Blueprint: placement-aggregate-generation