Add host info to instance action events
Currently, the admins can use instance action API to look up the
instance actions info, it is very useful for admins. But the
host info of events are not recorded and not exposed to user.

This host info helps admins to find the host of events
occurring, especially, when some failed events happened, the
admins can find the failed host and log as soon as possible.

We propose to add ``host`` field to
``InstanceActionEvent`` object and related API responses.

APIImpact

Co-Authored-By: Zhenyu Zheng <zhengzhenyu@huawei.com>

Change-Id: I8534bd637e19eda66a901444f36394794078926b
Blueprint: add-host-to-instance-action-events